feat: suppress reports from unexported interface methods (#52)
When an interface method is not exported, its only implementations will
be local to the module under analysis. Thus, the implementations can be
checked, and the caller can assume the error was wrapped.

This reasoning does not apply to exported interface methods, even if the
interface happens to also have an unexported method. (An implementation
can override just the public methods by embedding the interface or a
struct that implements it.)